Feat : 기능 추가, Body : 게시물,자산 api모듈 추가 , Paging 관련 내용 추가

+++ src/main/java/com/pingpong/jlab/pingpong/domain/record/service/RecordService.java
@@ -45,7 +45,7 @@ public class RecordService {
     public ApiResponse addUserRecord(RecordRequestDto recordInfo, String userinfo){
         
         Optional<User> user= userRepository.findByUserid(userinfo);
-        Asset asset = new Asset((Long)1,"cryptocurrency","bitcoin","65","BTCUSDT");
+        Asset asset = new Asset(1,"cryptocurrency","bitcoin","65","BTCUSDT");
         User recUser = user.get();
         recordInfo.setUser(recUser);
         recordInfo.setAsset(asset);
